Moved repositories to a new folder and removed unnecessary usings.
