+++ EyeCT4Rails/EyeCT4RailsASP/App_Start/BundleConfig.cs
@@ -1,5 +1,4 @@
-﻿using System.Web;
-using System.Web.Optimization;
+﻿using System.Web.Optimization;
 
 namespace EyeCT4RailsASP
 {
