+++ EyeCT4Rails/EyeCT4RailsASP/Controllers/CleanupController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Web.Mvc;
 using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
-using EyeCT4RailsLogic;
+using EyeCT4RailsLogic.Repositories;
 using static EyeCT4RailsASP.Models.UserUtilities;
 
 namespace EyeCT4RailsASP.Controllers
