+++ EyeCT4Rails/EyeCT4RailsASP/Controllers/RideController.cs
@@ -4,7 +4,7 @@ using System;
 using System.Web.Mvc;
 using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
-using EyeCT4RailsLogic;
+using EyeCT4RailsLogic.Repositories;
 using Newtonsoft.Json;
 
 namespace EyeCT4RailsASP.Controllers
