+++ EyeCT4Rails/EyeCT4RailsASP/Controllers/UserController.cs
@@ -4,7 +4,7 @@ using static EyeCT4RailsASP.Models.UserUtilities;
 using System.Web.Mvc;
 using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
-using EyeCT4RailsLogic;
+using EyeCT4RailsLogic.Repositories;
 
 namespace EyeCT4RailsASP.Controllers
 {
