+++ EyeCT4Rails/EyeCT4RailsASP/Models/UserUtilities.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Security.Cryptography;
-using System.Text;
-using System.Web.ModelBinding;
-using EyeCT4RailsLib.Classes;
+﻿using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
 
 namespace EyeCT4RailsASP.Models
