+++ EyeCT4Rails/EyeCT4RailsASP/Views/Cleanup/HistoryOfJob.cshtml
@@ -1,5 +1,4 @@
-﻿@using System.Web.UI.WebControls
-@using EyeCT4RailsLib.Classes
+﻿@using EyeCT4RailsLib.Classes
 
 @{
     ViewBag.Title = "Geschiedenis";
