+++ EyeCT4Rails/EyeCT4RailsLib/Classes/Function.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using EyeCT4RailsLib.Enums;
 
 namespace EyeCT4RailsLib.Classes
