+++ EyeCT4Rails/EyeCT4RailsLib/Enums/JobType.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace EyeCT4RailsLib.Enums
+﻿namespace EyeCT4RailsLib.Enums
 {
     public enum JobType
     {
