+++ EyeCT4Rails/EyeCT4RailsLib/Enums/TramType.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace EyeCT4RailsLib.Enums
 {
