+++ EyeCT4Rails/EyeCT4RailsLogic/Helpers/Hashing.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace EyeCT4RailsLogic.Helpers
 {
